third_party/WebKit/LayoutTests/fast/canvas/OffscreenCanvas-currentColor.html

<!DOCTYPE html>
<title>Test that the fill and stroke style attributes yield opaque black when set to 'currentcolor' on an OffscreenCanvasRenderingContext2D.</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(function() {
    var aCanvas = new OffscreenCanvas(50, 50);
    var ctx = aCanvas.getContext('2d');
    ctx.fillStyle = 'red';
    ctx.fillStyle = 'currentcolor'
    assert_equals(ctx.fillStyle, '#000000');

    ctx.strokeStyle = 'red';
    ctx.strokeStyle = 'currentcolor'
    assert_equals(ctx.strokeStyle, '#000000');
});
</script>
